pyproject.toml
[project]
name = "vibe-trading"
version = "0.1.0"
description = "Multi-agent cryptocurrency trading system using pi_agent_core"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Local packages
    "pi-agent-core",
    # Data sources
    "python-binance>=1.0.19",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    # Technical analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    # Storage
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    # Tools
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    # Memory
    "rank-bm25>=0.2.2",
    # HTTP
    "httpx>=0.26.0",
    # Logging
    "loguru>=0.7.2",
    "pyyaml>=6.0.3",
    "greenlet>=3.3.2",
    "openai>=2.30.0",
    "fastapi>=0.135.3",
    "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pi_agent_core", "src/pi_ai", "src/pi_logger", "src/vibe_trading"]

[tool.hatch.build.targets.sdist]
include = ["src/"]

[project.scripts]
vibe-trade = "vibe_trading.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
